+++ packages/components/form-list/index.ts
@@ -1,0 +1,8 @@
+import { withInstall } from "@viaz/utils";
+import FormList from "./src/index.vue";
+
+const VzFormList = withInstall(FormList);
+
+export { VzFormList };
+
+export default VzFormList;